requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
statsmodels==0.14.1
scikit-allel==1.3.7
scipy==1.13.1
fpdf==1.7.2
